# public/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'API G-10'
  description: "API de la SAE G-10. Pour l'utiliser, il faut se connecter avec un compte utilisateur et obtenir un token d'authentification. Ensuite, il faut ajouter ce token dans la section Authentification."
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
paths:
  /login:
    post:
      tags:
        - Authentification
      description: 'Connecte un utilisateur'
      operationId: 1da689c688cc418e88f9e9d668f01b16
      requestBody:
        description: "Les informations de connexion de l'utilisateur"
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  description: "Adresse email de l'utilisateur"
                  type: string
                password:
                  description: "Mot de passe de l'utilisateur"
                  type: string
              type: object
      responses:
        '200':
          description: 'Utilisateur connecté'
          content:
            application/json:
              schema:
                properties:
                  token: { description: "Jeton d'authentification JWT", type: string }
                type: object
        '401':
          description: 'Non autorisé'
  /register:
    post:
      tags:
        - Authentification
      description: 'Enregistre un nouvel utilisateur'
      operationId: c758f54bda9079f9ebe361693213d40c
      requestBody:
        description: "Les informations de connexion de l'utilisateur"
        required: true
        content:
          application/json:
            schema:
              required:
                - nom
                - prenom
                - adresse
                - code_postal
                - ville
                - email
                - password
              properties:
                nom:
                  description: "Nom de l'utilisateur"
                  type: string
                prenom:
                  description: "Prénom de l'utilisateur"
                  type: string
                adresse:
                  description: "Adresse de l'utilisateur"
                  type: string
                code_postal:
                  description: "Code postal de l'utilisateur"
                  type: string
                ville:
                  description: "Ville de l'utilisateur"
                  type: string
                email:
                  description: "Adresse email de l'utilisateur"
                  type: string
                password:
                  description: "Mot de passe de l'utilisateur"
                  type: string
              type: object
      responses:
        '200':
          description: 'Utilisateur créé'
        '401':
          description: 'Non autorisé'
  /user:
    get:
      tags:
        - Authentification
      description: "Se déconnecte l'utilisateur connecté"
      operationId: 7162e217cedb565bfc023a06d437ffa5
      responses:
        '200':
          description: 'Utilisateur connecté'
        '401':
          description: 'Non autorisé'
  /refresh:
    get:
      tags:
        - Authentification
      description: "Rafraîchit le token de l'utilisateur connecté"
      operationId: 2f339b146faea0a9f0245caef7d52107
      responses:
        '200':
          description: 'Token rafraîchi'
        '401':
          description: 'Non autorisé'
  /profil:
    get:
      tags:
        - Client
      description: 'Récupère le profil du client'
      operationId: 1825fa62241313dd09e00cd5bc4a9d14
      responses:
        '200':
          description: 'Profil du client'
        '404':
          description: 'Client non trouvé'
  /clients:
    get:
      tags:
        - Client
      description: 'Récupère la liste des clients'
      operationId: 62bbf3d7e2c130391994c9f22fe5534e
      parameters:
        -
          name: search
          in: query
          description: 'Rechercher par nom'
          required: false
          schema:
            type: string
        -
          name: sort
          in: query
          description: 'Trier par nom ou ville'
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  status: { type: string }
                  clients: { type: array, items: { $ref: '#/components/schemas/Client' } }
                type: object
        '404':
          description: 'Aucun client trouvé'
  '/clients/{id}':
    get:
      tags:
        - Client
      description: 'Récupère le client par son ID'
      operationId: cce02351b4e36b00e12f4503faa9d486
      parameters:
        -
          name: id
          in: path
          description: 'ID du client'
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  status: { type: string }
                  client: { $ref: '#/components/schemas/Client' }
                  user: { $ref: '#/components/schemas/User' }
                type: object
        '404':
          description: 'Client non trouvé'
    put:
      tags:
        - Client
      description: 'Mettre à jour le client'
      operationId: ed46ad7ebce2bffa44c3d50b5218f4de
      requestBody:
        description: 'Données à mettre à jour'
        required: true
        content:
          application/json:
            schema:
              properties:
                client:
                  $ref: '#/components/schemas/Client'
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  status: { type: string }
                  client: { $ref: '#/components/schemas/Client' }
                  user: { $ref: '#/components/schemas/User' }
                type: object
        '404':
          description: 'Client non trouvé'
...
components:
  schemas:
...
    Client:
      title: Client
      properties:
        id:
          description: 'Identifiant unique du client'
          type: integer
          format: int64
        user_id:
          description: "Identifiant de l'utilisateur du client"
          type: integer
        nom:
          description: 'Nom du client'
          type: string
        prenom:
          description: 'Prénom du client'
          type: string
        telephone:
          description: 'Numéro de téléphone du client'
          type: string
        adresse:
          description: 'Adresse du client'
          type: string
        code_postal:
          description: 'Code postal du client'
          type: string
        ville:
          description: 'Ville du client'
          type: string
        pays:
          description: 'Pays du client'
          type: string
      type: object
...
    User:
      title: User
      properties:
        id:
          description: "Identifiant unique de l'utilisateur"
          type: integer
          format: int64
        name:
          description: "Nom de l'utilisateur"
          type: string
        email:
          description: "Adresse email de l'utilisateur"
          type: string
        email_verified_at:
          description: "Date de vérification de l'adresse email"
          type: string
          format: date-time
        password:
          description: "Mot de passe de l'utilisateur"
          type: string
        remember_token:
          description: 'Jeton de connexion'
          type: string
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer